commit via shell script

+++ Project2/part3.c
@@ -41,13 +41,17 @@ void alarm_handler(int signal){
       running_child += 1;
     }
 
-    printf("ALARM: continuing child[%d]\n", pid_pool[running_child]);
-    kill(pid_pool[running_child], SIGCONT);
-
-    sleep(4);
-
     pid_t w;
     int wstatus;
+    if (w = waitpid(pid_pool[running_child], &wstatus, WNOHANG) != 0){
+      printf("NOPE\n");
+      flag_boi -= 1;
+    }
+    else{
+      printf("ALARM: continuing child[%d]\n", pid_pool[running_child]);
+      kill(pid_pool[running_child], SIGCONT);
+      sleep(4);
+    }
     if (w = waitpid(pid_pool[running_child], &wstatus, WNOHANG) != 0){
       printf("NOPE\n");
       flag_boi -= 1;
